:fire: factor task-sequence code out of cli.process
